Moves files into various folders and updates imports. Fixes height issue.
