+++ src/renderer/src/components/Onboarding/PrivacyPolicy/index.tsx
@@ -2,17 +2,17 @@ import React from 'react'
 import { styled } from '@mui/material/styles'
 import { useIntl } from 'react-intl'
 
-import { BLUE_GREY, DARK_GREY, VERY_LIGHT_GREY, WHITE } from '../../colors'
-import BarChart from '../../images/BarChart.svg'
-import BustInSilhouette from '../../images/BustInSilhouette.svg'
-import Wrench from '../../images/Wrench.svg'
-import ChevronUp from '../../images/chevrondown-expanded.svg'
-import ChevronDown from '../../images/chevrondown.svg'
-import ClosedLockWithKey from '../../images/closed_lock_with_key.png'
-import RaisedFistMediumSkinTone from '../../images/raised_fist_medium_skin_tone.png'
-import RaisedHandMediumSkinTone from '../../images/raised_hand_medium_skin_tone.png'
-import RedDot from '../../images/red_dot.png'
-import { Text } from '../Text'
+import { BLUE_GREY, DARK_GREY, VERY_LIGHT_GREY, WHITE } from '../../../colors'
+import BarChart from '../../../images/BarChart.svg'
+import BustInSilhouette from '../../../images/BustInSilhouette.svg'
+import Wrench from '../../../images/Wrench.svg'
+import ChevronUp from '../../../images/chevrondown-expanded.svg'
+import ChevronDown from '../../../images/chevrondown.svg'
+import ClosedLockWithKey from '../../../images/closed_lock_with_key.png'
+import RaisedFistMediumSkinTone from '../../../images/raised_fist_medium_skin_tone.png'
+import RaisedHandMediumSkinTone from '../../../images/raised_hand_medium_skin_tone.png'
+import RedDot from '../../../images/red_dot.png'
+import { Text } from '../../Text'
 import { DiagnosticItem } from './DiagnosticItem'
 import { MetricsDiagnosticsPermissionToggle } from './MetricsDiagnosticsPermissionToggle'
 import { PointContainer } from './PointContainer'
