+++ src/renderer/src/components/Onboarding/PrivacyPolicy/privacyPolicyMessages.ts
