+++ src/renderer/src/constants.ts
@@ -1,0 +1,4 @@
+export const DEVICE_NAME_MAX_LENGTH_GRAPHEMES = 60
+export const DEVICE_NAME_MAX_BYTES = 512
+export const PROJECT_NAME_MAX_LENGTH_GRAPHEMES = 100
+export const PROJECT_NAME_MAX_BYTES = 512
