+++ src/renderer/src/routes/Onboarding/CreateJoinProjectScreen.tsx
@@ -5,7 +5,7 @@ import { defineMessages, useIntl } from 'react-intl'
 
 import { BLACK, BLUE_GREY, DARK_GREY, WHITE } from '../../colors'
 import { Button } from '../../components/Button'
-import { OnboardingScreenLayout } from '../../components/OnboardingScreenLayout'
+import { OnboardingScreenLayout } from '../../components/Onboarding/OnboardingScreenLayout'
 import { Text } from '../../components/Text'
 import AddPersonImage from '../../images/add_person.png'
